pyproject.toml
[project]
name = "folio-builder-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apispec>=6.8.2",
    "apispec-webframeworks>=1.2.0",
    "cachetools==5.3.3",
    "cloudinary==1.38.0",
    "deprecated==1.2.14",
    "email-validator==2.2.0",
    "flask==3.0.1",
    "flask-admin==1.6.1",
    "flask-caching==2.3.0",
    "flask-cors==4.0.0",
    "flask-jwt-extended==4.6.0",
    "flask-limiter==3.5.1",
    "flask-login==0.6.3",
    "flask-mail==0.9.1",
    "flask-migrate==4.0.5",
    "flask-pydantic-spec>=0.8.6",
    "flask-sqlalchemy==3.1.1",
    "flask-swagger==0.2.14",
    "flask-swagger-ui==4.11.1",
    "flask-wtf==1.2.1",
    "gunicorn==21.2.0",
    "jinja2==3.1.4",
    "limits==3.13.0",
    "mako==1.3.5",
    "markdown-it-py==3.0.0",
    "openpyxl==3.1.5",
    "phonenumbers==8.13.45",
    "pillow==11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic==2.8.2",
    "pygments==2.18.0",
    "pyjwt<2.10",
    "pytest==8.2.2",
    "pytest-flask==1.3.0",
    "python-dotenv==1.0.1",
    "python-slugify==8.0.3",
    "requests==2.32.3",
    "sqlalchemy==2.0.31",
    "text-unidecode==1.3",
    "typing-extensions==4.12.2",
    "werkzeug==3.0.3",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingImports = "none"
reportMissingTypeStubs = "none"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]